+++ src/django_elasticsearch_dsl_drf/tests/__init__.py
@@ -1,3 +1,5 @@
 from .test_filtering import *
+from .test_helpers import *
 from .test_ordering import *
 from .test_search import *
+from .test_views import *