+++ src/django_elasticsearch_dsl_drf/views.py
@@ -1,11 +1,18 @@
+# -*- coding: utf-8 -*-
+
 """
 Base ViewSets.
 """
+from __future__ import absolute_import, unicode_literals
+
+from django.http import Http404
 
 from elasticsearch_dsl import Search
 from elasticsearch_dsl.connections import connections
 
 from rest_framework.viewsets import ReadOnlyModelViewSet
+
+from .utils import DictionaryProxy
 
 
 __title__ = 'django_elasticsearch_dsl_drf.views'
@@ -18,6 +25,7 @@ __all__ = ('BaseDocumentViewSet',)
 class BaseDocumentViewSet(ReadOnlyModelViewSet):
     """Base document viewset."""
 
+    document_uid_field = "id"
     document = None  # Re-define
 
     def __init__(self, *args, **kwargs):
@@ -32,3 +40,39 @@ class BaseDocumentViewSet(ReadOnlyModelViewSet):
     def get_queryset(self):
         """Get queryset."""
         return self.search.query()
+
+    def get_object(self):
+        """Get object."""
+        queryset = self.get_queryset()
+        lookup_url_kwarg = self.lookup_url_kwarg or self.lookup_field
+        if lookup_url_kwarg not in self.kwargs:
+            raise AttributeError(
+                "Expected view %s to be called with a URL keyword argument "
+                "named '%s'. Fix your URL conf, or set the `.lookup_field` "
+                "attribute on the view correctly." % (
+                    self.__class__.__name__,
+                    lookup_url_kwarg
+                )
+            )
+
+        if lookup_url_kwarg == 'id':
+            obj = self.document.get(id=self.kwargs[lookup_url_kwarg])
+            return DictionaryProxy(obj.to_dict())
+        else:
+            queryset = queryset.filter(
+                'term',
+                **{self.document_uid_field: self.kwargs[lookup_url_kwarg]}
+            )
+
+            count = queryset.count()
+            if count == 1:
+                obj = queryset.execute().hits.hits[0]['_source']
+                return DictionaryProxy(obj)
+
+            elif count > 1:
+                raise Http404(
+                    "Multiple results matches the given query. "
+                    "Expected a single result."
+                )
+
+            raise Http404("No result matches the given query.")