+++ CHANGELOG.rst
@@ -15,6 +15,13 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.1.8
+-----
+2017-07-26
+
+- Python2 fixes.
+- Documentation and example project improvements.
+
 0.1.7
 -----
 2017-07-25